requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
docker==7.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.9
python-git==2018.2.1
requests==2.32.3
Send2Trash==1.8.3
smmap==5.0.1
urllib3==2.2.3
